+++ packages/tools/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./extensions";
 export * from "./factories";
 export * from "./document";
-export { createOfficialToolCapabilitiesForDocument } from "./document/capabilities";
+export { createCapabilitiesForDocument } from "./document/capabilities";
 export {
   KIT_LEGACY_LAYER_PRESET,
   type KitLegacyLayerPreset,